Move InteropViewHolder to a separate file
